Add files via upload
I didn't know that I will be asked to upload all the work I have done. So the codes which didn't work I just removed completely from the file here is what I have done in the last two days since I was preparing for the final submission and considering the time shortage I just made a new file if the previous one was not working well.